--- codigo/Cliente.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Cliente
-{
-    internal class Cliente
-    {
-        private string nome;
-
-        public Cliente(string nome)
-        {
-            this.nome = nome;
-        }
-
-        static void Main(string[] args)
-        {
-
-        }
-    }
-}
