+++ codigo/staLuzia_Bulldogs/Cliente.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace staLuzia_Bulldogs
+{
+    internal class Cliente
+    {
+        private string nome;
+
+        public Cliente(string nome)
+        {
+            this.nome = nome;
+        }
+
+        static void Main(string[] args)
+        {
+
+        }
+    }
+}
