+++ codigo/staLuzia_Bulldogs/Mesa.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Mesa
+namespace staLuzia_Bulldogs
 {
     internal class Mesa
     {
