+++ codigo/staLuzia_Bulldogs/Restaurante.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Restaurante
+namespace staLuzia_Bulldogs
 {
     class Restaurante
     {
